spliting out AddComponentDialog class into its own file
